requirements.txt
openai==1.2.*
pytest==7.4.*
python-dotenv==1.0.*